DROOLS-1526: adding rule and compilation check to report empty decision table input entries
